+++ DiplomaProject.8392/services/Profile/Profile.Domain/Entities/City.cs
@@ -10,7 +10,7 @@ namespace Profile.Domain.Entities
     public class City: BaseEntity
     {
         public string Name { get; set; }
-        public int CountryId { get; set; }
+        public long CountryId { get; set; }
         public Country Country { get; set; }
         public ICollection<User> Users { get; set; }
     }
